add ultrasphinx support for query and ranged-query MVA types
MVA fields are integer array filters.  They work similar to the single
value filters.  For more info on MVA check the sphinx manual.

To use in ultrasphinx, add a :concatenate entry as normal, but replace
:association_sql with :association_table (the join table) and add :doc_id
as the column matching the current model id.
The purpose of this is to avoid the extra join of a giant join table.
Sphinx will run a query separately and merge the results into the index.

No delta conditions yet: the delta index will scan the full join table.

TODO: API should probably change, :doc_id isn't very intuitive
TODO: delta conditions on the mva queries and tests